Maybe fix messages, maybe work on SQLite
